Experiments in working with sbt-spark-packages
