--- src/main/python/.gitignore
@@ -1,5 +1,0 @@
-*.pyc
-/dist/
-/build/
-/*.egg-info
-
